+++ backend_laravel/app/Http/Controllers/EntrenamientoController.php
@@ -1,6 +1,4 @@
 <?php
-
-// app/Http/Controllers/ClaseController.php
 namespace App\Http\Controllers;
 
 use App\Http\Resources\EntrenamientoResource;
@@ -10,6 +8,9 @@ use App\Models\Entrenamiento;
 use App\Models\Entrenador;  
 use Illuminate\Http\Request;
 use Illuminate\Database\Eloquent\SoftDeletes;
+use App\Policies\EntrenamientoPolicy;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\Log;
 
 class EntrenamientoController extends Controller
 {
@@ -21,9 +22,9 @@ class EntrenamientoController extends Controller
     }
 
     // Obtener una clase específica por su ID
-    public function show($id)
+    public function show($slug)
     {
-        $entrenamiento = Entrenamiento::with('deporte')->find($id);
+        $entrenamiento = Entrenamiento::where('slug', $slug)->firstOrFail();
         if (!$entrenamiento) {
             return response()->json(['error' => 'Entrenamiento no encontrado'], 404);
         }
@@ -33,6 +34,11 @@ class EntrenamientoController extends Controller
     // Crear una nueva clase
     public function store(Request $request)
     {
+        $entrenador = auth('entrenador')->user();
+        
+        if (!$entrenador) {
+            return response()->json(['error' => 'No autenticado'], 401);
+        }
         $request->validate([
             'nombre' => 'required|string|max:255',
             'descripcion' => 'required|string|max:500',
@@ -49,36 +55,47 @@ class EntrenamientoController extends Controller
             'max_plazas' => $request->max_plazas,
             'precio' => $request->precio,
             'deporte_id' => $request->deporte_id,
+            'entrenador_id'=>$entrenador->id,
         ]);
 
         return response()->json($entrenamiento, 201);
     }
 
     // Actualizar una clase existente
-    public function update(Request $request, $id)
-    {
-        $entrenamiento = Entrenamiento::find($id);
+    public function update(Request $request, $slug)
+    {  
+        Log::debug('Token recibido: ' . request()->bearerToken());
+        $entrenador = auth('entrenador')->user();
+        $entrenamiento = Entrenamiento::where('slug', $slug)->firstOrFail();
         if (!$entrenamiento) {
-            return response()->json(['error' => 'entrenamiento no encontrada'], 404);
+            Log::debug('entrenamiento no encontrado');
+            return response()->json(['error' => 'entrenamiento no encontrado'], 404);
         }
-
+        if (!$entrenador->can('update', $entrenamiento)) {
+            Log::debug('no autorizado');
+            return response()->json(['error' => 'No autorizado'], 403);
+        }
+        Log::debug('llegue aqui');
         $request->validate([
-            'nombre' => 'required|string|max:255',
-            'deportes_id' => 'required|exists:deportes,deportes_id',  
+            'nombre' => 'nullable|string|max:255',
+            'deportes_id' => 'nullable|exists:deportes,id',  
         ]);
-
+        Log::debug('llegue aqui tambien');
         $entrenamiento->update($request->all());
         return response()->json($entrenamiento);
     }
 
     // Eliminar una clase
-    public function destroy($id)
+    public function destroy($slug)
     {
-        $entrenamiento = Entrenamiento::find($id);
+        $entrenador = auth('entrenador')->user();
+        $entrenamiento = Entrenamiento::where('slug', $slug)->firstOrFail();
         if (!$entrenamiento) {
             return response()->json(['error' => 'Clase no encontrada'], 404);
         }
-
+        if (!$entrenador->can('delete', $entrenamiento)) {
+            return response()->json(['error' => 'No autorizado'], 403);
+        }
         $entrenamiento->delete();
         return response()->json(['message' => 'Clase eliminada']);
     }
